### 2023-07-13 #5656 @editorialbot
[PRE REVIEW]: chessboard: An R package for creating network connections based on chess moves
**Submitting author:** @ahasverus (<a href="http://orcid.org/0000-0002-5537-5294">Nicolas Casajus</a>)
**Repository:** https://github.com/FRBCesab/chessboard
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1
