+++ joss-paper/paper.bib
@@ -33,7 +33,7 @@
   year    = {2006},
   volume  = {196},
   pages   = {493},
-  doi     = {0.1016/j.ecolmodel.2006.02.015}
+  doi     = {10.1016/j.ecolmodel.2006.02.015}
 }
 
 @book{legendre-2012,
