+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   chess-pieces: chess-pieces.html
   chessboard: chessboard.html
   visualization-tools: visualization-tools.html
-last_built: 2023-07-13T08:08Z
+last_built: 2023-07-13T14:19Z
 
