### 2023-08-02 #5656 @Nikoleta-v3
👋🏻 @GisKid, @Pentaonia, @ssujit would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "chessboard: An R package for creating network connections based on chess moves". You can find more information at the top of this Github issue 🆙
