### 2023-08-06 #5656 @Pentaonia
Hey @Nikoleta-v3 and @ahasverus,

as described in my profile, I am currently finishing my Master's degree. So I don't have much experience in reviewing. This would be my first review for this journal.

The topic is quite interesting, as I have been teaching about spatial statistics in the last semester and how to apply methods that use these kinds of networks!
However, if my level of experience is enough for you, I would be happy to join, help and learn!

All the best
Louis
