### 2023-08-08 #5656 @Nikoleta-v3
Hey @Pentaonia 👋🏻, thank you for replying.

 I see no issue with you being an MSc student! In my experience, first-time reviewers always provide very detailed reviews. And in case you have any issues with the review, you can always ping me, and I can potentially be of help.
