### 2023-08-09 #5656 @Pentaonia
Hey @Nikoleta-v3,

perfect, then I will start as soon as possible.

So when I start, can I just put the checklist ([here](https://joss.readthedocs.io/en/latest/review_checklist.html)) in a new comment of mine and tick off the list items or comment below when I am done or have found something?
