### 2023-08-09 #5656 @gsapijaszko
> 👋🏻 @jedalong and @gsapijaszko would any of you be willing to review this submission for JOSS?
> [...]
> Please let me know if you're available 😄 Thank you!

Yes, let's try :).

Regards,
Grzegorz
