### 2023-08-10 #5656 @Nikoleta-v3
Hey @Pentaonia 👋🏻

> So when I start, can I just put the checklist ([here](https://joss.readthedocs.io/en/latest/review_checklist.html)) in a new comment of mine and tick off the list items or comment below when I am done or have found something?

Correct! Just no here, as not on this issue. This is the pr- review issue. I am going to start the **review** issue now since we have two reviewers!
