### 2023-08-10 FRBCesab/chessboard#2 @Pentaonia
JOSS Review from me [done]
**Review link:** openjournals/joss-reviews#5753 (hopefully its now linked correctly)

**Missing:**
- [Writing Contributor Guidelines + _maybe_ an issue template](https://github.com/FRBCesab/chessboard/community)
- You mention that the output is compatible with other R packages. This has been mentioned for ggplot2 and sf. I think it would be a helpful addition to include at least one or two functions from the other packages as well. That way newbies would have a starting point.
- A link to the "reference" page that points to for example [adespatial](https://search.r-project.org/CRAN/refmans/adespatial/html/aem.html) regarding the MEM/AEM maps.
- Indicate where the 'real world example' comes from. Whether it is part of a real analysis (might a citation be appropriate?) or a "fictional" example of what can be done.

**Hints:**
- It might be a good idea to include an extra chunk in the "Getting Started" vignette where you load additional packages (like ggplot2, sf). As you did for the visualisation tools page. In a vanilla R environment, at least I had to load the "ggplot2" package in addition.
- Maybe you can adapt the functions on the "Chess pieces" page to work with the data created/used on the "Getting Started" page or the one from the paper.

**Question:**
- Since you use functions from the tidyverse, why don't you work with tibbles? Because of compatibility reasons?
- Have you checked the plots on other devices? Idk if its my fault, but it looks slightly different than in your visualization vignette. In my case, it is not a perfect circle around the focus node and it is slightly offset. It is not that dramatic but I wanted to mention it (in [this](https://frbcesab.github.io/chessboard/articles/chessboard.html#edge-list) step):
```
gg_chessboard(nodes) +
  geom_node(nodes, focus = "2-3")
```
![image](https://github.com/FRBCesab/chessboard/assets/93387223/a7d7373e-37e5-4485-9627-eedb039ff5e7)

**Manuscript**
- Lines 9/10, suggestion: "It can handle directed (asymmetric) and undirected (symmetric) spatial (or non-spatial) network connections."
- Lines 29/30: Might be the same case as for lines 9/10.

**Thumbs up:**
- Great package! I would definitely incorporate it into my lecture.
- I really like the 'reference' page in your vignette.
