### 2023-08-13 eed4355
doc: add contributor guidelines - fix #4
