### 2023-08-13 02ae996
doc: add links to mem/aem man pages - fix #5
