### 2023-08-13 7995c9e
doc: add setup chunk in vignette - fix #6
