### 2023-08-13 1424caa
joss: fix typos in paper - fix #7
