+++ README.Rmd
@@ -42,8 +42,8 @@ chessboard <img src="man/figures/package-sticker.png" align="right" style="float
 ## Overview
 
 The package `chessboard` provides functions to work with **directed**
-(asymmetric) and **undirected** (symmetrical) spatial (or not) **networks**. It
-implements different methods to detect neighbors, all based on the chess game
+(asymmetric) and **undirected** (symmetric) spatial (or non-spatial) **networks**. 
+It implements different methods to detect neighbors, all based on the chess game
 (it goes beyond the rook and the queen) to create complex connectivity scenarios.
 
 `chessboard` can handle spatial networks, but it does not explicitly use geographical
