+++ vignettes/chessboard.Rmd
@@ -17,8 +17,8 @@ knitr::opts_chunk$set(collapse  = TRUE,
 
 
 The aim of the package `chessboard` is to provide tools to work with **directed**
-(asymmetric) and **undirected** (symmetrical) spatial (or not) **networks**. It
-implements different methods to detect neighbors, all based on the chess game
+(asymmetric) and **undirected** (symmetrical) spatial (or non-spatial) **networks**. 
+It implements different methods to detect neighbors, all based on the chess game
 (it goes beyond the rook and the queen available in many R packages) to create 
 complex connectivity scenarios.
 
