+++ DESCRIPTION
@@ -9,7 +9,7 @@ Authors@R: c(
            email   = "nicolas.casajus@fondationbiodiversite.fr",
            comment = c(ORCID = "0000-0002-5537-5294")))
 Description: Provides functions to work with directed (asymmetric) and 
-    undirected (symmetrical) spatial networks. It makes easier the creation of 
+    undirected (symmetric) spatial networks. It makes easier the creation of 
     connectivity matrix, i.e. a binary matrix of dimensions n x n, where n is 
     the number of nodes (sampling units) indicating the presence (1) or the 
     absence (0) of an edge (link) between pairs of nodes. Different network 
