### 2023-08-13 5926cd8
doc: update package documentation
