+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   chess-pieces: chess-pieces.html
   chessboard: chessboard.html
   visualization-tools: visualization-tools.html
-last_built: 2023-08-13T11:02Z
+last_built: 2023-08-13T11:06Z
 
