### 2023-08-13 2b8bca8
doc: modify header levels (CONTRIBUTING.md)
