### 2023-08-13 FRBCesab/chessboard#2 @ahasverus
Hi @Pentaonia :wave:

Thank you for reviewing this paper and the associated repository.

Here is my answers to your different points:

1. **Missing**

- **Adding Contributor Guidelines**

> Thanks for the suggestion! Done in #4. The markdown file can be found [here](https://github.com/FRBCesab/chessboard/blob/main/CONTRIBUTING.md).

- **Adding an Issue template**

> Thanks for the suggestion! Done in #3. I have added three templates: one for [reporting bugs](https://github.com/FRBCesab/chessboard/blob/main/.github/ISSUE_TEMPLATE/bug_report.md), one for [requesting features](https://github.com/FRBCesab/chessboard/blob/main/.github/ISSUE_TEMPLATE/feature_request.md), and one for [other questions](https://github.com/FRBCesab/chessboard/blob/main/.github/ISSUE_TEMPLATE/other_issue.md).

- **Adding examples to link to other packages**

> Some examples on how to link `chessboard` to `ggplot2` and `sf` are already included in the [Get started](https://frbcesab.github.io/chessboard/articles/chessboard.html) vignette, especially at the end of section [edge list](https://frbcesab.github.io/chessboard/articles/chessboard.html#edge-list). Do you think is enough or do you had something different in mind?

- **Adding links to reference pages of MEM/AEM**

> Thanks for the suggestion! Done in #5.

- **Real world example**

> You are right! Data used in the [Get started](https://frbcesab.github.io/chessboard/articles/chessboard.html) vignette are fictional but inspired from a real study (not published yet). As these data are fictional maybe I could use _fictional example_ instead of `real-world example`. What do you think?

2. **Hints**

- **Loading additional packages**

> Done in #6. Thanks for the suggestion!

- **Using data of "Get started" in "Chess pieces" vignette**

> The objective of the [Chess pieces](https://frbcesab.github.io/chessboard/articles/chess-pieces.html) vignette is to illustrate the use of the different arguments of the functions `fool()`, `pawn()`, `rook()`, etc. To do that I needed a square sampling design (in that case 9 transects x 9 quadrats). I decided to use a different sampling design in the Get started vignette just to illustrate the fact that `chessboard` works with a non-square sampling design. For the manuscript I decided to use a smaller square sampling design for visibility reasons and to provide a third example. Do you think this will confuse the user?

3. **Questions**

- **Why don't you work with tibbles?**

> There is no particular reason why I use `data.frame` instead of `tibble`. It's just a matter of habit and also to usually limit external dependencies (which is not a good reason here ;)).

- **Plot rendering**

> I have tested the package on different OS and I never saw this kind of rendering issues. What OS are you using? Do you use RStudio? Is the screenshot coming from RStudio or from an export as PNG/JPG? I will investigate this issue but to me is not link to `chessboard`.

4. **Manuscript**

- Suggestion for lines 9/10 and lines 29/30

> Done in #7. I also fixed these typos in the comment of the chunk at lines 49/50 of the manuscript, and in the `README`, the `DESCRIPTION` file, and in the `Get started` vignette.

Thank again for your review and contribution!
