### 2023-08-13 #5753 @ahasverus
Hi @gsapijaszko and @Pentaonia :wave:
Thank you for agreeing to review our works!

@Pentaonia: Thanks for your review. I have answered your questions in the issue [https://github.com/FRBCesab/chessboard/issues/2](https://github.com/FRBCesab/chessboard/issues/2).
