### 2023-08-14 FRBCesab/chessboard#2 @Pentaonia
Hi @ahasverus  🖖

Thank you for your answers!
I just took a glimpse into your modified documents: looks great! Here my answers to the remaining questions:

1. Missing

- Adding examples to link to other packages

> Some examples on how to link chessboard to ggplot2 and sf are already included in the [Get started](https://frbcesab.github.io/chessboard/articles/chessboard.html) vignette, especially at the end of section [edge list](https://frbcesab.github.io/chessboard/articles/chessboard.html#edge-list). Do you think is enough or do you had something different in mind?

>> So I had in mind that you might also include one or two functions from the other packages you mention in the paper. So that the end user clearly knows how to import your edge list/connectivity matrix (or maybe another data.frame/matrix) into spdep or igraph. I mean, you might have a clear idea of where the links are and how to build them. Someone else who is not so deeply involved in the field might not see the links directly and especially not the corresponding functions.

- Real world example

> You are right! Data used in the [Get started](https://frbcesab.github.io/chessboard/articles/chessboard.html) vignette are fictional but inspired from a real study (not published yet). As these data are fictional maybe I could use fictional example instead of real-world example. What do you think?

>> Yeah it would be a bit clearer, or something like this: "Fictional scenario with real-world parallels".

2. Hints

- Using data of "Get started" in "Chess pieces" vignette

> The objective of the [Chess pieces](https://frbcesab.github.io/chessboard/articles/chess-pieces.html) vignette is to illustrate the use of the different arguments of the functions fool(), pawn(), rook(), etc. To do that I needed a square sampling design (in that case 9 transects x 9 quadrats). I decided to use a different sampling design in the Get started vignette just to illustrate the fact that chessboard works with a non-square sampling design. For the manuscript I decided to use a smaller square sampling design for visibility reasons and to provide a third example. Do you think this will confuse the user?

>> Puh, that's a good question. I would say that it is a matter of judgement: Between a larger grid of 9x9 instead of 5x5 in the paper or the possibility of not having to introduce anything new in the chess pieces vignette to be able to include the functions in an example.
>> If you want to leave it as it is in the paper, which is no problem, then (1) maybe refer to it and "expand" the grid from the paper to explain the functions or (2) introduce it as a new data set/grid in the chess pieces vignette. This might introduce something kind of new again, but I would assume that the reader has no problem with a new data set being used on a new page of the vignette. Especially if it is as simple as a grid, without much spatial information.

3. Questions

- Why don't you work with tibbles?
> There is no particular reason why I use data.frame instead of tibble. It's just a matter of habit and also to usually limit external dependencies (which is not a good reason here ;)).
>>  I was just curious because you use a handful of functions from the tidyverse, but yeah, never mind! No reason to criticise anything here! ;)

- Plot rendering
> I have tested the package on different OS and I never saw this kind of rendering issues. What OS are you using? Do you use RStudio? Is the screenshot coming from RStudio or from an export as PNG/JPG? I will investigate this issue but to me is not link to chessboard.
>> I'm using Windows. I have checked in all three environments (Vanilla R, R Studio and my always used VS Code whats opening the `x11()` plot window). I am using R version 4.3.1. Export as PDF works as expected, but as jpg/png, same as above... but then it seems like a "me problem"! ;)

All the best!
