### 2023-08-28 dec2ed5
doc: update list of authors in metadata (#8)
