+++ README.md
@@ -135,8 +135,9 @@ A companion paper has been submitted to [Journal of Open Source
 Software](https://joss.theoj.org/). In the meantime, if you want to use
 the package, please cite it as:
 
-> Casajus N (2023) chessboard: An R package for creating network
-> connections based on chess moves. R package version 0.1. URL:
+> Casajus N, Rievrs Borges E, Tabacchi E, Fried G & Mouquet N (2023)
+> chessboard: An R package for creating network connections based on
+> chess moves. R package version 0.1. URL:
 > <https://github.com/frbcesab/chessboard>.
 
 ## Contributing
