+++ README.Rmd
@@ -144,7 +144,7 @@ vignette describes the plotting functions available in `chessboard`
 A companion paper has been submitted to [Journal of Open Source Software](https://joss.theoj.org/).
 In the meantime, if you want to use the package, please cite it as: 
 
-> Casajus N (`r format(Sys.Date(), "%Y")`) chessboard: An R 
+> Casajus N, Rievrs Borges E, Tabacchi E, Fried G & Mouquet N (`r format(Sys.Date(), "%Y")`) chessboard: An R 
 package for creating network connections based on chess moves. 
 R package version 0.1. URL: <https://github.com/frbcesab/chessboard>.
 
