+++ DESCRIPTION
@@ -7,7 +7,27 @@ Authors@R: c(
            family  = "Casajus",
            role    = c("aut", "cre", "cph"),
            email   = "nicolas.casajus@fondationbiodiversite.fr",
-           comment = c(ORCID = "0000-0002-5537-5294")))
+           comment = c(ORCID = "0000-0002-5537-5294")),
+    person(given   = "Erica",
+           family  = "Rievrs Borges",
+           role    = "aut",
+           email   = "erica.rievrs@fondationbiodiversite.fr",
+           comment = c(ORCID = "0000-0001-7751-6265")),
+    person(given   = "Eric",
+           family  = "Tabacchi",
+           role    = "aut",
+           email   = "eric.tabacchi@univ-tlse3.fr",
+           comment = c(ORCID = "0000-0001-7729-4439")),
+    person(given   = "Guillaume",
+           family  = "Fried",
+           role    = "aut",
+           email   = "guillaume.fried@anses.fr",
+           comment = c(ORCID = "0000-0002-3653-195X")),
+    person(given   = "Nicolas",
+           family  = "Mouquet",
+           role    = "aut",
+           email   = "nicolas.mouquet@cnrs.fr",
+           comment = c(ORCID = "0000-0003-1840-6984")))
 Description: Provides functions to work with directed (asymmetric) and 
     undirected (symmetric) spatial networks. It makes easier the creation of 
     connectivity matrix, i.e. a binary matrix of dimensions n x n, where n is 
