### 2023-08-28 3f28f80
Update CITATION.cff
