+++ CITATION.cff
@@ -9,7 +9,7 @@ type: software
 license: GPL-2.0-or-later
 title: 'chessboard: Create Network Connections based on Chess Moves'
 version: '0.1'
-abstract: 'Provides functions to work with directed (asymmetric) and undirected (symmetrical)
+abstract: 'Provides functions to work with directed (asymmetric) and undirected (symmetric)
   spatial networks. It makes easier the creation of connectivity matrix, i.e. a binary
   matrix of dimensions n x n, where n is the number of nodes (sampling units) indicating
   the presence (1) or the absence (0) of an edge (link) between pairs of nodes. Different
@@ -20,12 +20,32 @@ authors:
   given-names: Nicolas
   email: nicolas.casajus@fondationbiodiversite.fr
   orcid: https://orcid.org/0000-0002-5537-5294
+- family-names: Rievrs Borges
+  given-names: Erica
+  email: erica.rievrs@fondationbiodiversite.fr
+  orcid: https://orcid.org/0000-0001-7751-6265
+- family-names: Tabacchi
+  given-names: Eric
+  email: eric.tabacchi@univ-tlse3.fr
+  orcid: https://orcid.org/0000-0001-7729-4439
+- family-names: Fried
+  given-names: Guillaume
+  email: guillaume.fried@anses.fr
+  orcid: https://orcid.org/0000-0002-3653-195X
+- family-names: Mouquet
+  given-names: Nicolas
+  email: nicolas.mouquet@cnrs.fr
+  orcid: https://orcid.org/0000-0003-1840-6984
 preferred-citation:
   type: manual
   title: 'chessboard: An R package for creating network connections based on chess
     moves'
   authors:
   - name: Casajus Nicolas
+  - name: Rievrs-Borges Erica
+  - name: Tabacchi Eric
+  - name: Fried Guillaume
+  - name: Mouquet Nicolas
   year: '2023'
   notes: R package version 0.1
   url: https://github.com/frbcesab/chessboard
