### 2023-08-30 #5753 @Pentaonia
Hey @Nikoleta-v3 im almost done with the review, my checklist has been ticked off. Im just waiting for an answer to the comment https://github.com/FRBCesab/chessboard/issues/2#issuecomment-1678096341. After that, I think there is nothing to stop it being published.
