### 2023-09-05 FRBCesab/chessboard#2 @ahasverus
Hi @Pentaonia ,

I'm just back from vacation. Thank you for your answers! Here is my answer to the remaining points:

- **Adding examples to link to other packages**

> So I had in mind that you might also include one or two functions from the other packages you mention in the paper. So that the end user clearly knows how to import your edge list/connectivity matrix (or maybe another data.frame/matrix) into spdep or igraph. I mean, you might have a clear idea of where the links are and how to build them. Someone else who is not so deeply involved in the field might not see the links directly and especially not the corresponding functions.

>> I have created a new issue (#9) that will show small examples to extend `chessboard` w/ different packages. I agree w/ you this can be a very good added value.

- **Real world example**

> Yeah it would be a bit clearer, or something like this: "Fictional scenario with real-world parallels". (But it's late now so this suggestion might be nonsense.)

>> I agree maybe it's too late for this suggestion.

- **Using data of "Get started" in "Chess pieces" vignette**

> Puh, that's a good question. I would say that it is a matter of judgement: Between a larger grid of 9x9 instead of 5x5 in the paper or the possibility of not having to introduce anything new in the chess pieces vignette to be able to include the functions in an example. If you want to leave it as it is in the paper, which is no problem, then (1) maybe refer to it and "expand" the grid from the paper to explain the functions or (2) introduce it as a new data set/grid in the chess pieces vignette. This might introduce something kind of new again, but I would assume that the reader has no problem with a new data set being used on a new page of the vignette. Especially if it is as simple as a grid, without much spatial information.

>> In my opinion, the user/reader can easily understand these two different datasets without being lost. But I agree and I will add a sentence in the [Chess pieces](https://frbcesab.github.io/chessboard/articles/chess-pieces.html) vignette indicating that the network used in this vignette is derived from the one used in the paper.

Again: thanks for your contribution.
