### 2023-09-06 185a127
doc: explain dataset in vignette (#10)
