### 2023-09-06 0074f74
doc: section extending chessboars (wip)
