### 2023-09-06 7a0a96e
doc: extending w/ ggplot2
