### 2023-09-06 2fe27a2
doc: extending w/ igraph
