+++ DESCRIPTION
@@ -48,6 +48,7 @@ Imports:
     tidyr
 Suggests: 
     knitr,
+    igraph,
     patchwork,
     rmarkdown,
     testthat (>= 3.0.0)
