### 2023-09-06 e2c1c04
Update CITATION.cff
