### 2023-09-06 1167edd
Deploying to gh-pages from @ FRBCesab/chessboard@2fe27a26e994c2af11c6e5323391d9d4fd2fe648 🚀
