### 2023-09-07 8f7e9f4
doc: review get started vignette
