### 2023-09-07 3c72674
doc: review get started vignette
