### 2023-09-07 a6a02db
doc: add extending chessboard - fix #9
