### 2023-09-07 4cd4797
Deploying to gh-pages from @ FRBCesab/chessboard@a6a02db5a59a1e1361497fd8f3eb796fca8999f7 🚀
