### 2023-09-07 FRBCesab/chessboard#2 @ahasverus
I think I have answered to all your comments.
All the bests!
