### 2023-09-19 FRBCesab/chessboard#2 @Pentaonia
Hey @ahasverus,

looks great to me!
