### 2023-09-19 #5753 @Pentaonia
Hey @Nikoleta-v3 @ahasverus,

I just checked the changes and closed the issue (https://github.com/FRBCesab/chessboard/issues/2#issuecomment-1725364846). From my point of view it looks great. I have nothing to comment on.

All the best
Louis
