### 2023-10-04 #5753 @Nikoleta-v3
Hello @ahasverus  and thank you for your patience! I can confirm that I now have had the time to review the manuscript, documentation, and go over the source code. The package is well-developed, tested, and documented. I am very happy to move forward with acceptance. I just have some minor comments.

Regarding the paper, in lines 19-20 where you describe the connectivity matrix, use a math environment. So, $n \times n$, and in line 20, simply use $n$.

You state: "chessboard implements the following rules to detect neighbors and to create edges:

- Degree
- Orientation
- Direction

In Figure 1, you give an example of the neighbors; however, these concepts are not clearly captured in the figure. In my understanding the argument `degree` is for the degree, `reverse` and `directed` are for the direction, and orientation is for values like "Pawn," "Fool," etc. Could you update the figure to better reflect this?

Also, for the direction parameter, there are two input arguments,`reverse` and `directed`. Both of these arguments can be set to `True` at the same time, and then the package just uses reverse. Is this behavior documented somewhere in the function? Do you want to include a warning message if a user sets both as `True`?

Regarding the contributing file, thank you very much for adding one and thanks to our reviewers. The "Contributing code" section needs a bit more information and examples. For instance, including a command for the  "Fork this repository" would be helpful. Additionally, running the test suite is not trivial for someone new to software development and open source, but just reading the contributing file 😄  Please assume that someone new to this might want to contribute to the package and provide a bite more guidance.
