### 2023-10-09 FRBCesab/chessboard#11 @ahasverus
Use math environment in paper.md
As suggested in [5753](https://github.com/openjournals/joss-reviews/issues/5753#issuecomment-1746804452), need to to use mathematical environment in `paper.md` when describing dimensions of the connectivity matrix (lines 19-20).
