### 2023-10-09 FRBCesab/chessboard#12 @ahasverus
Improve Fig. 1 in paper.md
As mentionned in [5753](https://github.com/openjournals/joss-reviews/issues/5753#issuecomment-1746804452), need to improve Fig. 1 in `paper.md` to better explain the rules implemented in the package to detect neighbors and their relation with the arguments method, degree, directed, reverse.
