### 2023-10-09 FRBCesab/chessboard#13 @ahasverus
Improve Contributor guidelines
As mentioned in [5753](https://github.com/openjournals/joss-reviews/issues/5753#issuecomment-1746804452), need to improve the section `Contributing code` of the `CONTRIBUTING.md` file
