### 2023-10-09 4963b92
joss: use math environment - fix #11
