### 2023-10-09 0e8081a
doc: improve contribute code - fix #13
