### 2023-10-09 a7adb58
Deploying to gh-pages from @ FRBCesab/chessboard@0e8081a9465b876fbf27619fe1b54377e5ca1bcc 🚀
