### 2023-10-09 60b5ac4
joss: improve fig 1 - fix #12
