### 2023-10-09 #5753 @ahasverus
> Hello @ahasverus and thank you for your patience! I can confirm that I now have had the time to review the manuscript, documentation, and go over the source code. The package is well-developed, tested, and documented. I am very happy to move forward with acceptance. I just have some minor comments.

Hi @Nikoleta-v3! Thank you for your positive feedbacks.

> Regarding the paper, in lines 19-20 where you describe the connectivity matrix, use a math environment. So, n×n, and in line 20, simply use n.

This has been done in [#4963b92](https://github.com/FRBCesab/chessboard/commit/4963b922cfc4d6c6880eb083fe49dcb65c0577f3)

> You state: "chessboard implements the following rules to detect neighbors and to create edges:
>
> * Degree
> * Orientation
> * Direction
>
> In Figure 1, you give an example of the neighbors; however, these concepts are not clearly captured in the figure. In my understanding the argument `degree` is for the degree, `reverse` and `directed` are for the direction, and orientation is for values like "Pawn," "Fool," etc. Could you update the figure to better reflect this?

I have made some modifications to Fig.1 in [#60b5ac4](https://github.com/FRBCesab/chessboard/commit/60b5ac48bf090b5ddd72de31c3810e331ae37a41). What do you think?

> Also, for the direction parameter, there are two input arguments,`reverse` and `directed`. Both of these arguments can be set to `True` at the same time, and then the package just uses reverse. Is this behavior documented somewhere in the function? Do you want to include a warning message if a user sets both as `True`?

Indeed. Both `directed` and `reverse` arguments control the direction of the network. The argument `reverse` can be used to change the default direction of the network. But if the argument `directed` is set to `FALSE` the argument `reverse` is ignored as it is mentioned in the documentation. For instance [here](https://frbcesab.github.io/chessboard/reference/create_edge_list.html).

> Regarding the contributing file, thank you very much for adding one and thanks to our reviewers. The "Contributing code" section needs a bit more information and examples. For instance, including a command for the "Fork this repository" would be helpful. Additionally, running the test suite is not trivial for someone new to software development and open source, but just reading the contributing file 😄 Please assume that someone new to this might want to contribute to the package and provide a bite more guidance.

Thanks for this suggestion: I have improved the section _Contribute Code_ of the **Contributing Guidelines** in [#0e8081a](https://github.com/FRBCesab/chessboard/commit/0e8081a9465b876fbf27619fe1b54377e5ca1bcc)
