### 2023-10-09 e224f63
Deploying to gh-pages from @ FRBCesab/chessboard@60b5ac48bf090b5ddd72de31c3810e331ae37a41 🚀
