### 2023-10-09 #5753 @Nikoleta-v3
> Indeed. Both directed and reverse arguments control the direction of the network. The argument reverse can be used to change the default direction of the network. But if the argument directed is set to FALSE the argument reverse is ignored as it is mentioned in the documentation. For instance [here](https://frbcesab.github.io/chessboard/reference/create_edge_list.html).

Perfect! Thank you and thank you for addressing my comments.

At this point could @ahasverus you please:

- [ ] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release 0.1.0
 Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
