### 2023-10-10 f0c36e9
Deploying to gh-pages from @ FRBCesab/chessboard@60b5ac48bf090b5ddd72de31c3810e331ae37a41 🚀
