### 2023-10-10 a0b4665
Deploying to gh-pages from @ FRBCesab/chessboard@60b5ac48bf090b5ddd72de31c3810e331ae37a41 🚀
