### 2023-10-10 #5753 @ahasverus
Thank you so much @Nikoleta-v3!

- [x] Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release 0.1.0

> Here is the link to the Release **v0.1.0** on GitHub: <https://github.com/FRBCesab/chessboard/releases/tag/v0.1.0>

- [x] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

> Here is the URL to associated **Zenodo repository**: <https://zenodo.org/record/8424609>

- [x] Please list the DOI of the archived version here

> The **DOI** provided by Zenodo is: [10.5281/zenodo.8424609](https://doi.org/10.5281/zenodo.8424609)
