### 2023-10-10 414cd38
doc: add JOSS citation
