### 2023-10-10 0b19c7b
doc: fix URL issues
