### 2023-10-10 e3b65f2
Update CITATION.cff
