+++ CITATION.cff
@@ -36,19 +36,6 @@ authors:
   given-names: Nicolas
   email: nicolas.mouquet@cnrs.fr
   orcid: https://orcid.org/0000-0003-1840-6984
-preferred-citation:
-  type: manual
-  title: 'chessboard: An R package for creating network connections based on chess
-    moves'
-  authors:
-  - name: Casajus Nicolas
-  - name: Rievrs-Borges Erica
-  - name: Tabacchi Eric
-  - name: Fried Guillaume
-  - name: Mouquet Nicolas
-  year: '2023'
-  notes: R package version 0.1
-  url: https://github.com/frbcesab/chessboard
 repository-code: https://github.com/frbcesab/chessboard
 url: https://github.com/frbcesab/chessboard
 contact:
