### 2023-10-10 86d1944
ci: remove render-readme action
