### 2023-10-10 43f7b1b
Update CITATION.cff
