+++ README.Rmd
@@ -87,9 +87,9 @@ vignette for further details.
 
 Finally, the package can also produce objects that will be used later in
 Moran's Eigenvector Maps 
-([MEM](https://search.r-project.org/CRAN/refmans/adespatial/html/mem.html), 
+([MEM](https://rdrr.io/cran/adespatial/man/mem.html), 
 Dray _et al._ 2006) and Asymetric Eigenvector Maps 
-([AEM](https://search.r-project.org/CRAN/refmans/adespatial/html/aem.html),
+([AEM](https://rdrr.io/cran/adespatial/man/aem.html),
 Blanchet _et al._ 2008), methods available in the package
 [`adespatial`](https://cran.r-project.org/package=adespatial) (Dray _et al._ 2023):
 
