### 2023-10-11 19cd57a
cran: remove links to adespatial
