+++ R/edges_weights_vector.R
@@ -1,12 +1,13 @@
 #' Create an edges weights vector
 #' 
 #' @description
-#' Creates an edges weights vector that can be used in [adespatial::aem()]. 
-#' Resulting edges weights equal to 0 will be replaced by `4 x max(w)`, 
-#' where `max(w)` is the maximal weight in the matrix.
+#' Creates an edges weights vector that can be used in `aem()` of the package 
+#' `adespatial`. Resulting edges weights equal to 0 will be replaced by 
+#' `4 x max(w)`, where `max(w)` is the maximal weight in the matrix.
 #'
 #' @param x a `list` of length 2. The nodes_by_edges matrix returned by 
-#'   `nodes_by_edges_matrix()` or [adespatial::aem.build.binary()].
+#'   `nodes_by_edges_matrix()` (or `aem.build.binary()` of the package 
+#'   `adespatial`).
 #'   
 #' @param y a `data.frame` with the following three columns: `from` 
 #'   (the first node of the edge), `to` (the second node of the edge), and 
@@ -68,7 +69,7 @@ edges_weights_vector <- function(x, y) {
   
   if (any(!(names(x) %in% c("se.mat", "edges")))) {
     stop("Argument 'x' must be a nodes-by-edges matrix as returned by ",
-         "nodes_by_edges_matrix() or adespatial::aem.build.binary()", 
+         "nodes_by_edges_matrix() or aem.build.binary() (adespatial)", 
          call. = FALSE)
   }
   
