+++ man/chessboard-package.Rd
@@ -4,7 +4,7 @@
 \name{chessboard-package}
 \alias{chessboard}
 \alias{chessboard-package}
-\title{chessboard: Create Network Connections based on Chess Moves}
+\title{chessboard: Create Network Connections Based on Chess Moves}
 \description{
 Provides functions to work with directed (asymmetric) and undirected (symmetric) spatial networks. It makes easier the creation of connectivity matrix, i.e. a binary matrix of dimensions n x n, where n is the number of nodes (sampling units) indicating the presence (1) or the absence (0) of an edge (link) between pairs of nodes. Different network objects can be produced by chessboard: node list, neighbor list, edge list, connectivity matrix.
 }
