### 2023-10-11 537ffbf
Deploying to gh-pages from @ FRBCesab/chessboard@19cd57ac2c2c9dd4e773057e35062f1f2cfaf212 🚀
