### 2023-10-11 2bfbcd9
Deploying to gh-pages from @ FRBCesab/chessboard@257480140fc63c1f57a65ef62687c8048804e7d5 🚀
