### 2023-10-13 #5753 @danielskatz
I've suggested some minor changes in https://github.com/FRBCesab/chessboard/pull/14 - please merge this, or let me know what you disagree with, then we can proceed.
