### 2023-10-13 1d995d5
Deploying to gh-pages from @ FRBCesab/chessboard@f2943afbc4c6b43d26eeff8d6fa1a385034f4950 🚀
