### 2023-10-13 #5753 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
