### 2023-10-13 #5753 @danielskatz
Congratulations to @ahasverus (Nicolas Casajus) and co-authors on your publication!!

JOSS depends on volunteers and wouldn't be successful without you
