### 2023-10-13 2c41146
Deploying to gh-pages from @ FRBCesab/chessboard@71eb728372e8d8ff4715c8ee36f7c9451a16a7dc 🚀
